# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/dash/article/add": {
            "post": {
                "description": "Crée un article avec ses informations et une image associée",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "Ajouter un nouvel article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nom de l'article",
                        "name": "nom",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Description de l'article",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "description": "Prix de l'article",
                        "name": "prix",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Quantité en stock",
                        "name": "stock",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID du commerçant",
                        "name": "commercant_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID de la catégorie",
                        "name": "categorie_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Largeur de l'image",
                        "name": "largeur",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Hauteur de l'image",
                        "name": "hauteur",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Ordre de l'image",
                        "name": "ordre",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Type de l'image (jpg, png, etc.)",
                        "name": "type",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Taille de l'image",
                        "name": "taille",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Image encodée en base64",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article créé avec succès",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Requête invalide ou image manquante",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Erreur serveur lors de l'ajout de l'article",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/dash/article/list": {
            "get": {
                "description": "Retourne une liste paginée d'images (5 par page).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "Récupérer les images d'articles",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Numéro de la page (commence à 1)",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ArticleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/dash/article/{id}": {
            "get": {
                "description": "Retourne les informations détaillées d'un article (images, catégorie, commerçant)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "Récupérer les détails d'un article",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de l'article",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ArticleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/dash/article/{id}/delete": {
            "delete": {
                "description": "Supprime un article en fonction de son ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "Supprimer un article",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de l'article",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article supprimé avec succès",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ID invalide",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Article introuvable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entities.ArticleImage": {
            "type": "object",
            "properties": {
                "article_id": {
                    "type": "integer"
                },
                "hauteur": {
                    "type": "integer"
                },
                "image_id": {
                    "type": "integer"
                },
                "largeur": {
                    "type": "integer"
                },
                "order": {
                    "type": "integer"
                },
                "taille": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "entities.ArticleResponse": {
            "type": "object",
            "properties": {
                "categorie": {
                    "$ref": "#/definitions/entities.Categorie"
                },
                "commercant": {
                    "$ref": "#/definitions/entities.Commercant"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.ArticleImage"
                    }
                },
                "nom": {
                    "type": "string"
                },
                "prix": {
                    "type": "number"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "entities.Categorie": {
            "type": "object",
            "properties": {
                "categorie_id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "nom": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                }
            }
        },
        "entities.Commercant": {
            "type": "object",
            "properties": {
                "adresse": {
                    "type": "string"
                },
                "commercant_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "nom": {
                    "type": "string"
                },
                "telephone": {
                    "type": "string"
                }
            }
        },
...
    }
}